# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
servers:
    - url: https://api.memos.com
paths:
...
    /v1/classes:
        get:
            tags:
                - ClassService
            description: ListClasses lists classes based on filters.
            operationId: ClassService_ListClasses
            parameters:
                - name: filter
                  in: query
                  description: Optional filter string.
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  description: Optional page size.
                  schema:
                    type: integer
                    format: int32
                - name: pageToken
                  in: query
                  description: Optional page token.
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListClassesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - ClassService
            description: CreateClass creates a new class.
            operationId: ClassService_CreateClass
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateClassRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Class'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        Class:
            required:
                - uid
                - displayName
                - visibility
            type: object
            properties:
                name:
                    type: string
                    description: |-
                        The resource name of the class.
                         Format: classes/{class}
                uid:
                    type: string
                    description: The unique identifier (slug) of the class.
                displayName:
                    type: string
                    description: The display name of the class.
                description:
                    type: string
                    description: Optional description of the class.
                creator:
                    readOnly: true
                    type: string
                    description: |-
                        The resource name of the creator.
                         Format: users/{user}
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    description: Output only. The last update timestamp.
                    format: date-time
                visibility:
                    enum:
                        - CLASS_VISIBILITY_UNSPECIFIED
                        - CLASS_PUBLIC
                        - CLASS_PROTECTED
                        - CLASS_PRIVATE
                    type: string
                    description: The visibility of the class.
                    format: enum
                inviteCode:
                    type: string
                    description: Optional invite code for joining the class.
                settings:
                    allOf:
                        - $ref: '#/components/schemas/ClassSettings'
                    description: Optional class settings.
            description: Class represents a classroom for error book.
        ClassSettings:
            type: object
            properties:
                studentMemoVisibility:
                    type: boolean
                    description: Whether students can see each other's memos.
                allowAnonymous:
                    type: boolean
                    description: Whether to allow anonymous submissions.
                defaultStudentVisibility:
                    enum:
                        - CLASS_VISIBILITY_UNSPECIFIED
                        - CLASS_PUBLIC
                        - CLASS_PROTECTED
                        - CLASS_PRIVATE
                    type: string
                    description: Default visibility for student memos.
                    format: enum
                enableTagTemplates:
                    type: boolean
                    description: Whether to enable class tag templates.
                maxMembers:
                    type: integer
                    description: Maximum number of members allowed (0 for unlimited).
                    format: int32
                requireMemberApproval:
                    type: boolean
                    description: Whether to require approval for new members.
            description: Class settings for configuration.
        CreateClassRequest:
            required:
                - class
            type: object
            properties:
                classId:
                    type: string
                    description: Optional class ID (slug). If not provided, will be generated.
                class:
                    allOf:
                        - $ref: '#/components/schemas/Class'
                    description: The class to create.
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListClassesResponse:
            type: object
            properties:
                classes:
                    type: array
                    items:
                        $ref: '#/components/schemas/Class'
                    description: The list of classes.
                nextPageToken:
                    type: string
                    description: Next page token.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...